requirements.txt
langchain
langchain-community
python-dotenv
pypdf
langchain_text_splitters
langchain_core
langchain_huggingface
faiss-cpu
langchain-groq
fastapi
langserve
uvicorn
sse_starlette
streamlit